Improve naming of v14+ instructions/types
Give meaningful names to the instructions introduced in ain v14+. These
are just names I made up based on what the instructions appear to do.

The iter<...> type is renamed to wrap<...> since it seems to be some
kind of generic wrapper type, not specific to the foreach syntax.

The enum#82 type is renamed to option<...>. This appears to be a 2-value
type where the first value is some wrapped value, and the second is a
code indicating sucess or failure of the operation that produced it.

The unknown_100 type is renamed to iwrap<...>. This type is used
whenever an interface type is wrapped in a wrap<...>. This is probably
just to distinguish it from a regular struct type, since interface types
are represented by 2 values.