Treat ain v11+ 'rank' as boolean
In ain v11+, the 'rank' field of a type is always 1 or 0. If it's 1,
then a sub-type should be read. There is no need to consult the data
type to determine the existence of a sub-type, as was done previously.

ain_open will now raise an error if this field is non-boolean in v11+.