ain/ex: transcoding support
Add ain_read_conv and ex_read_conv for transcoding ain/ex files at
read-time. These functions take an additional argument which should be a
pointer to a function that will convert strings to an alternative text
encoding.

This is mainly intended for the alice-tools GUI, where transcoding
on-demand is challenging due to memory management issues. In the future,
it would be good for the alice-tools CLI and xsystem4 to use UTF-8
internally as well.