Add archive_release_file, archive_copy_descriptor
archive_release_file is the opposite operation to archive_load_file: it
releases the loaded data and returns the descriptor to the unloaded
state.

archive_copy_descriptor copies a descriptor to be stored for later use.
This can be used together with archive_for_each to create an external
index.