Allow building as a shared library
libsys4 can now be built as a shared library by setting the
'default_library' meson option to either 'shared' or 'both'. E.g.

    meson build -Ddefault_library=both

This is needed for sys4c because dune expects a dynamic library to exist
when installing for some reason.

Also, 'install : true' is now set on the library to prevent mes from
building it as a thin archive. This causes errors in the sys4c build
because it is necessary to move the archive as part of the build
(breaking the relative paths stored in a thin archive).