Improve ALD header un-obfuscation
This fixes https://github.com/kichikuou/xsystem4-android/issues/1.

In the downloaded editions of some games, the first three bytes of ALD
are obfuscated. Before this patch, only the download edition of Sengoku
Rance was confirmed to work. This patch adds support for ALDS in the
download edition of GALZOO Island and Choukou Sennin Haruka.

- Haruka uses the same magic number as Sengoku Rance, but the existing
  code did not work when the second byte was not 0x4c (i.e. when ALD has
  more entries than 21845).
- GALZOO uses a different magic number.