ald: Magic number detection
ALD archives in downloadable edition of some games are obfuscated by
adding a fixed value to the first 3 bytes of the archive. Since this
magic number varies from game to game, this change scans the pointer
table to determine the correct value of the first 3 bytes, and then
calculates the magic number from there.