Fix bug in `_ald_get`
This function does not work correctly at high optimization levels. The
incorrect behavior can be reproduced in xsystem4 by starting a new game
in Rance VI and using Sill's healing skill in the initial battle.

The game will check if a file exists in Rance6WA.ald with index 3132
(3131), and `_ald_get` reports that it does. Then when the game tries to
load the sound effect, `_ald_get` correctly reports that it does not
exist, at which point the game calls `system.Error`.

This is "fixed" by disabling optimization of this function. I was not
able to track down the root cause of the error. asan/ubsan don't produce
any leads, debug builds don't reproduce the bug, and printing from
within this function causes the bug to disappear.